requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiogram_dialog==2.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.7.0
appdirs==1.4.4
APScheduler==3.11.0
attrs==24.3.0
betterLogger==1.7
betterlogging==1.0.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
dnspython==2.7.0
filelock==3.18.0
flake8==7.1.1
frozenlist==1.5.0
fsspec==2025.5.0
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.2.2
gspread==6.0.2
gspread_asyncio==2.0.0
h11==0.14.0
hf-xet==1.1.2
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.31.4
idna==3.10
Jinja2==3.1.5
joblib==1.5.0
magic-filter==1.0.12
MarkupSafe==3.0.2
mccabe==0.7.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
prettytable==3.16.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
pymongo==4.13.0
pyparsing==3.2.0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
StrEnum==0.4.15
sympy==1.14.0
telegram==0.0.1
texttable==1.7.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.3
triton==3.3.0
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
yarl==1.18.3
